refactor: consolidate removeAnnotationRecursive into AnnotationTrait
Both CleanUnusedComponents and PathFilter had identical local
implementations. Move the shared method into AnnotationTrait, replacing
the old collectAnnotations/traverseAnnotations/removeAnnotation methods
that are no longer used.
