+++ src/Processors/CleanUnusedComponents.php
@@ -15,6 +15,8 @@ use OpenApi\Generator;
  */
 class CleanUnusedComponents
 {
+    use Concerns\AnnotationTrait;
+
     protected bool $enabled;
 
     public function __construct(bool $enabled = false)
@@ -105,28 +107,5 @@ class CleanUnusedComponents
         }
 
         return [] !== $unusedComponents;
-    }
-
-    /**
-     * Recursively remove an annotation and all its nested children from the analysis.
-     */
-    protected function removeAnnotationRecursive(Analysis $analysis, OA\AbstractAnnotation $annotation): void
-    {
-        $analysis->removeAnnotation($annotation);
-
-        foreach (get_object_vars($annotation) as $property => $value) {
-            if (in_array($property, $annotation::$_blacklist)) {
-                continue;
-            }
-            if (is_array($value)) {
-                foreach ($value as $item) {
-                    if ($item instanceof OA\AbstractAnnotation) {
-                        $this->removeAnnotationRecursive($analysis, $item);
-                    }
-                }
-            } elseif ($value instanceof OA\AbstractAnnotation) {
-                $this->removeAnnotationRecursive($analysis, $value);
-            }
-        }
     }
 }
