+++ src/Processors/PathFilter.php
@@ -7,7 +7,6 @@
 namespace OpenApi\Processors;
 
 use OpenApi\Analysis;
-use OpenApi\Annotations as OA;
 use OpenApi\Generator;
 
 /**
@@ -19,6 +18,8 @@ use OpenApi\Generator;
  */
 class PathFilter
 {
+    use Concerns\AnnotationTrait;
+
     protected array $tags;
 
     protected array $paths;
@@ -97,26 +98,6 @@ class PathFilter
             }
 
             $analysis->openapi->paths = $filtered;
-        }
-    }
-
-    protected function removeAnnotationRecursive(Analysis $analysis, OA\AbstractAnnotation $annotation): void
-    {
-        $analysis->removeAnnotation($annotation);
-
-        foreach (get_object_vars($annotation) as $property => $value) {
-            if (in_array($property, $annotation::$_blacklist)) {
-                continue;
-            }
-            if (is_array($value)) {
-                foreach ($value as $item) {
-                    if ($item instanceof OA\AbstractAnnotation) {
-                        $this->removeAnnotationRecursive($analysis, $item);
-                    }
-                }
-            } elseif ($value instanceof OA\AbstractAnnotation) {
-                $this->removeAnnotationRecursive($analysis, $value);
-            }
         }
     }
 }
